+++ src/client/components/delete_btn/delete_btn.js
@@ -13,7 +13,7 @@ import './delete_btn.html';
 Template.delete_btn.helpers({
     // whether we show the disabled button
     enabledClass(){
-        return this.table.opt( 'deleteButtonEnabled', true, this.item ) || !Tabular._conf.hideDisabled ? '' : 'ui-transparent';
+        return this.table.opt( 'deleteButtonEnabled', true, this.item ) || !Tabular.configure().hideDisabled ? '' : 'ui-transparent';
     },
 
     // whether the displayed row is deletable ? defaulting to true