+++ src/client/components/edit_btn/edit_btn.js
@@ -13,7 +13,7 @@ import './edit_btn.html';
 Template.edit_btn.helpers({
     // whether to show the disabled button ?
     enabledClass(){
-        return this.table.opt( 'editButtonEnabled', true, this.item ) || !Tabular._conf.hideDisabled ? '' : 'ui-transparent';
+        return this.table.opt( 'editButtonEnabled', true, this.item ) || !Tabular.configure().hideDisabled ? '' : 'ui-transparent';
     },
 
     // whether the displayed row is deletable ? defaulting to true